fix(db): 股票域实体↔DDL 类型对齐 + v200 补表 + 工作流诊断增强
- 股票域 6 处布尔语义列实体 bool→i32（decision_validations/price_alerts/
  quant_runs/quant_strategies），守 PG 布尔约定（列 INTEGER、禁 bool）
- stock_analyses.decision_expected_holding_days Option<u32>→Option<i64>，
  适配上游 BIGINT 列（DDL 不动）
- 新增 v200_axinvest_stock_tables.rs：14 张 AxInvest 独有股票表
  (PHASE 1b CREATE TABLE IF NOT EXISTS + PHASE 2 索引)，修复全新库缺表，
  使「实体 vs 列」可静态审计一致
- harness/workflow_types StepProgressEvent 增 error 字段，
  rt-workflow engine 配套写入，诊断数据工具节点失败
- runtime-core/feature_flags 换 std::sync::RwLock，修嵌套 tokio runtime panic
- stock-analysis 回测/复盘/hit_rate 逻辑随类型修正同步
- 同步上游 master：workflow_executions.total_time_ms 守 BIGINT（i64），
  db_type_audit.py 统一为上游 527 行标准版

验证：cargo check -p axagent 通过；dprint/rustfmt 合规；
v100 total_time_ms 维持 BIGINT 与上游逐字一致

+++ src-tauri/crates/entities/src/stock_analyses.rs
@@ -27,7 +27,7 @@ pub struct Model {
     /// 时间维度: "ultra_short" | "short" | "mid" | "long"
     pub decision_time_horizon: Option<String>,
     /// 期望持有天数（交易日）
-    pub decision_expected_holding_days: Option<u32>,
+    pub decision_expected_holding_days: Option<i64>,
     /// 决策所用 LLM 的版本标识（用于复现实验）
     pub model_version: Option<String>,
     /// 关联到 L2 disk-cache 的快照 ID
